feat: add a shutdown hook to flush all unsent events
when the ReportingWorker starts, it will set a shutdown hook.
this shutdown hook will attempt to flush all unsent `eventsToSend` and `errors`.

also extracted some of the `drain` logic to a separate function to allow reuse by the shutdown hook which doesn't care about heartbeats and managing the worker's state, since it runs in a shutdown context.