test: testing the shutdown hook's functionality
asserting the agent sends a last beacon of events prior to shutting down.
this is literally a copy-paste of the `e2e/test-reload.py` test, with a much, much higher reportIntervalMs to make sure the agent doesn't report naturally.
the test will end the demo app, which will result in a shutdown hook triggering.